requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
pydantic>=2.5.0
python-dateutil>=2.8.2
gitpython>=3.1.40
